Updating Ubuntu Version to Fix glibc vuln
EVEC-89 #comment this will completely redownload and rebuild the image with a newer version of the ubuntu base image that has version "2.19-0ubuntu6.7" of the libc6 package. According to http://www.ubuntu.com/usn/usn-2900-1/ this fixes the vulnerability. (Note that the vulnerability is only fixed inside the Docker image. Further checks are needed for the host machine)